perf: 优化验证码验证和菜单数据结构
- 将验证码验证的 X 和 Y坐标类型从 int64改为 int
- 修正菜单结构中的 ApiList 字段名称为 MenuApi
- 优化滑动验证码验证逻辑

+++ server/internal/http/admin/ctl/auth_ctl.go
@@ -36,8 +36,8 @@ func (c *AuthCtl) Login(w http.ResponseWriter, r *http.Request) {
 		CaptchaKey string `json:"captcha_key" validate:"required"`
 		Md5        bool   `json:"md5"`
 		Point      struct {
-			X int64 `json:"x"`
-			Y int64 `json:"y"`
+			X int `json:"x"`
+			Y int `json:"y"`
 		}
 	}
